requirements.txt
rav
djangorestframework
django-cors-headers
dj-rest-auth
dj-rest-auth[with_social]
Pillow